requirements.txt
streamlit
pandas
openai
PyPDF2
python-docx
python-dotenv
plotly